Revert "Make asm jump labels local"
This reverts commit 935b39f2385c096dbfe6fa27121b2f1501211fe8.

.L L L$ prefixes are neither necessary nor desirable for assembly
source code, because assembly code labels are always of internal
linkage (static), and do not cause conflicts with other compilation
units (object files), unless .global is used to make them global, and
they appear in more than one object file.

Assembly kernel labels do not need to be made into "Local Symbols",
which are symbols which are totally omitted from debugging and symbol
table output *, and hence are unable to be used as breakpoint names,
or even seen as descriptive label names during step-by-step debugging,
or profiled as being the locations of hotspots.

(* Unless "as -L" is used.)

"Local Labels" use only digits, which are even harder to read. The
main motivation for local labels is during m4 macro expansion, when a
snippet of code needs a label whose relative position in the code is
always forward or backward from all branches to it, and hence it can
be done such as:

define(`macro1',`
 beqz loop_counter, 1f
 nop
1:
')

or

define(`macro2',`
1:
 nop
 bnez loop_counter, 1b
')

macro1
macro2    ; no conflict in label 1 -- it gets reused

The assembler automatically turns these numbered "Local Labels" into
regularly-named "local symbols" like label_1, label_2, etc., which are
then optionally outputted in the object file if -L is specified.  The
numbers can be reused later with no ambiguity, because the forward and
backward branches and the label form a local range.  It only makes
sense in m4 macros, where the "local label" is "local to the macro".

If removal of all non-global symbols from an object file is desired
because it is unwanted debugging infomation that is making the object
file too large, then you can use "strip -x" instead, without having to
use weird naming conventions which are not standard across object file
formats, and which make the source code harder to read and maintain.

Using periods at the beginning of labels to make them "Local Symbols"
confuses some syntax highlighting and auto-indenting editors, because
it looks like a pseudo-op, and since a pseudo-op must appear in the
second column or later, the editor might automatically indent it if
the first character typed on the line is '.'. If you really wanted a
label, you will have to manually un-indent it, since labels must start
in the first column.

The mangling of symbol names for locality conventions is the job of
the assembler or object code generator, not the assembly code author.

References:

https://sourceware.org/binutils/docs/as/Symbol-Names.html
https://sourceware.org/binutils/docs-2.18/as/Symbol-Names.html
https://stackoverflow.com/questions/51150860/assembly-label-prefixes